change optimizer base for RecomputeOptimizer
