+++ python/paddle/distributed/fleet/meta_optimizers/recompute_optimizer.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 
-from paddle.fluid.optimizer import RecomputeOptimizer as RO
+from paddle.incubate.optimizer import RecomputeOptimizer as RO
 
 from .meta_optimizer_base import MetaOptimizerBase
