+++ test/legacy_test/test_optimizer.py
@@ -656,8 +656,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -687,8 +689,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([b1_out])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -719,8 +723,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([b1_out.name])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -751,8 +757,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([mul_out, b2_out])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -783,8 +791,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([mul_out, b1_out])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -814,8 +824,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([b2_out, mul_out])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -846,8 +858,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([mul_x, b2_out])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -874,8 +888,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
 
     def test_apply_gradients(self):
         mul_out, b1_out, b2_out, mean_out = self.net()
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([b1_out])
         # apply backward
         params_grads = recompute_optimizer.backward(
@@ -912,8 +928,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
 
     def test_load(self):
         mul_out, b1_out, b2_out, mean_out = self.net()
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([b1_out])
         try:
             state_dict = {}
@@ -935,8 +953,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
             [op.type for op in mean_out.block.ops],
             ["mul", "dropout", "elementwise_add", "elementwise_add", "mean"],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([b1_out])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -980,8 +1000,10 @@ class TestRecomputeOptimizer(unittest.TestCase):
                 "mean",
             ],
         )
-        sgd_optimizer = optimizer.SGD(learning_rate=1.0)
-        recompute_optimizer = optimizer.RecomputeOptimizer(sgd_optimizer)
+        sgd_optimizer = paddle.optimizer.SGD(learning_rate=1.0)
+        recompute_optimizer = paddle.incubate.optimizer.RecomputeOptimizer(
+            sgd_optimizer
+        )
         recompute_optimizer._set_checkpoints([b1_out])
         opts, params_grads = recompute_optimizer.minimize(mean_out)
 
@@ -1050,8 +1072,8 @@ class TestRecomputeOptimizer(unittest.TestCase):
                     name="y", shape=[-1, 1], dtype='int64'
                 )
                 drop_res, prediction, cost = mlp(input_x, input_y)
-                sgd = fluid.optimizer.Adam(learning_rate=0.01)
-                sgd = fluid.optimizer.RecomputeOptimizer(sgd)
+                sgd = paddle.optimizer.Adam(learning_rate=0.01)
+                sgd = paddle.incubate.optimizer.RecomputeOptimizer(sgd)
                 sgd._set_checkpoints([prediction])
                 sgd.minimize(cost)
 
@@ -1115,8 +1137,8 @@ class TestRecomputeOptimizerCUDA(unittest.TestCase):
                     name="y", shape=[-1, 1], dtype='int64'
                 )
                 drop_res, prediction, cost = mlp(input_x, input_y)
-                sgd = fluid.optimizer.Adam(learning_rate=0.01)
-                sgd = fluid.optimizer.RecomputeOptimizer(sgd)
+                sgd = paddle.optimizer.Adam(learning_rate=0.01)
+                sgd = paddle.incubate.optimizer.RecomputeOptimizer(sgd)
                 sgd._set_checkpoints([prediction])
                 sgd.minimize(cost)
